server/requirements.txt
PyMuPDF
google-generativeai
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers